requirements.txt
together 
streamlit 
pandas 
matplotlib 
seaborn 
PyPDF2 
python-docx 
Pillow 
python-dotenv 
pyngrok 
PyMuPDF 
pytesseract